+++ java-r5rcore/src/org/ipea/r5r/R5RCore.java
@@ -48,6 +48,8 @@ public class R5RCore {
 
     private final String dataPath;
 
+    public final RDataFrame gtfsErrors;
+
     public double getWalkSpeed() {
         return this.routingProperties.walkSpeed;
     }
@@ -248,13 +250,18 @@ public class R5RCore {
         WorkerComponents.fileStorage = new R5RFileStorage(null);
 
         nativeElevationFunction = nativeElevationFunction.toUpperCase();
-        NetworkBuilder.useNativeElevation = !nativeElevationFunction.equals("NONE");
-        NetworkBuilder.elevationCostFunction = nativeElevationFunction;
+
+        NetworkBuilder builder = new NetworkBuilder();
+
+        builder.useNativeElevation = !nativeElevationFunction.equals("NONE");
+        builder.elevationCostFunction = nativeElevationFunction;
 
         dataPath = dataFolder;
         Path path = Paths.get(dataFolder).toAbsolutePath().normalize();
-        TransportNetwork network = NetworkBuilder.checkAndLoadR5Network(path.toString());
-        this.routingProperties = new RoutingProperties(network);
+        // network will be null if there were high priority errors
+        TransportNetwork network = builder.checkAndLoadR5Network(path.toString());
+        this.routingProperties = network == null ? null : new RoutingProperties(network);
+        this.gtfsErrors = builder.gtfsErrors;
     }
 
     // ---------------------------------------------------------------------------------------------------