+++ r-package/R/accessibility.R
@@ -55,7 +55,7 @@
 #'
 #' data_path <- system.file("extdata/poa", package = "r5r")
 #' r5r_network <- build_network(data_path)
-#' points <- read.csv(file.path(data_path, "poa_hexgrid.csv"))[1:5, ]
+#' points <- read.csv(file.path(data_path, "poa_hexgrid.csv"))[1:500, ]
 #'
 #' departure_datetime <- as.POSIXct(
 #'   "13-05-2019 14:00:00",